Split module into multiple submodules. Hide classes that shouldn't be public. Document classes.
